+++ src/CodeWF.AvaloniaControls.Showcase/Views/TestNoneWindowDemos/CodeWFWindowStandardDemo.axaml.cs
@@ -1,3 +1,4 @@
+using Avalonia.Controls.Templates;
 using CodeWF.AvaloniaControls.Controls;
 
 namespace CodeWF.AvaloniaControls.Showcase.Views.TestNoneWindowDemos;
@@ -7,5 +8,13 @@ public partial class CodeWFWindowStandardDemo : CodeWFWindow
     public CodeWFWindowStandardDemo()
     {
         InitializeComponent();
+        ApplyTitleBarSlotTemplates();
+    }
+
+    private void ApplyTitleBarSlotTemplates()
+    {
+        LeftContentTemplate = (IDataTemplate)Resources["SampleLeftContentTemplate"]!;
+        TitleBarContentTemplate = (IDataTemplate)Resources["SampleTitleBarContentTemplate"]!;
+        RightContentTemplate = (IDataTemplate)Resources["SampleRightContentTemplate"]!;
     }
 }